tests: Extract namespace helpers from utils
