+++ tests/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "io_utils.go",
         "job.go",
         "migration.go",
+        "namespace.go",
         "pod_servers.go",
         "serviceaccount.go",
         "utils.go",
